requirements.txt
openai-whisper
resemblyzer
numpy
soundfile
pyaudio
scipy
torch
keyboard
fastapi 
uvicorn 
python-multipart